+++ metrics-scala/src/test/scala/nl/grons/metrics4/scala/HistogramSpec.scala
@@ -16,13 +16,12 @@
 
 package nl.grons.metrics4.scala
 
-import org.mockito.Mockito.{when, verify}
+import org.mockito.IdiomaticMockito._
 import org.scalatest.Matchers._
-import org.scalatest.mockito.MockitoSugar._
-import org.scalatest.FunSpec
 import org.scalatest.OneInstancePerTest
+import org.scalatest.funspec.AnyFunSpec
 
-class HistogramSpec extends FunSpec with OneInstancePerTest {
+class HistogramSpec extends AnyFunSpec with OneInstancePerTest {
   describe("A histogram") {
     val metric = mock[com.codahale.metrics.Histogram]
     val histogram = new Histogram(metric)
@@ -30,19 +29,19 @@ class HistogramSpec extends FunSpec with OneInstancePerTest {
     it("updates the underlying histogram with an int") {
       histogram += 12
 
-      verify(metric).update(12)
+      metric.update(12) was called
     }
 
     it("updates the underlying histogram with a long") {
       histogram += 12L
 
-      verify(metric).update(12L)
+      metric.update(12L) was called
     }
 
     it("retrieves a snapshot for statistics") {
       val snapshot = mock[com.codahale.metrics.Snapshot]
-      when(snapshot.getMax).thenReturn(1L)
-      when(metric.getSnapshot).thenReturn(snapshot)
+      snapshot.getMax shouldReturn 1L
+      metric.getSnapshot shouldReturn snapshot
 
       histogram.max should equal (1L)
     }